move XdsLoadBalancer.XdsConfig to XdsLoadBalancerProvider
